Ползунок тона: тон Double, ширина CGFloat — компилятор постарше не выбирал между ними
Пятый и последний из этой цепочки. Положение бегунка считалось умножением
CGFloat-ширины на Double-тон, и дальше по цепочке всякое деление становилось
неоднозначным: у CGFloat и Double одинаковые операторы, а выбрать между ними
старому компилятору нечем. Здешний справляется, поэтому все пять дефектов
и дожили до публикации.

Шапка окна «О программе» заодно получила затемнение между сценой и надписью:
на снимке видно, что имя и подпись тонули в самой яркой части картинки —
тень вокруг букв с этим не справляется, потому что спорит не с фоном, а со светом.

+++ Sources/SRWLEDMenuBar/HueSlider.swift
@@ -21,7 +21,11 @@ struct HueSlider: View {
     var body: some View {
         GeometryReader { geometry in
             let width = max(1, geometry.size.width)
-            let knobX = width * shown
+            // Тон — Double, ширина — CGFloat. Пока их перемножали как есть,
+            // компилятор постарше объявлял неоднозначным всякое деление ниже
+            // по цепочке: у CGFloat и Double одинаковые операторы, и выбрать
+            // между ними ему было нечем.
+            let knobX = width * CGFloat(shown)
 
             ZStack(alignment: .leading) {
                 Capsule()
